Finished Gaussian2D implementation
Performance is good, accuracy is good, but it just doesn't seem to make much of a difference to the partial CMDs. It is not the significant difference I thought it would be. Probably just going to keep this code here; might not end up pushing it to the main branch. The Gaussian2D can be transplanted into the photometry package, though.