new mdf fitting procedure, remove spgbox
By using the same logarithmic variable transformation as was necessary for HMC we can greatly aid convergence of Optim.BFGS method for very fast maximum likelihood estimate. We can also use the inverse hessian appoximation built by BFGS to estimate standard errors which compare very favorably with the results from HMC. The new optimization is also not particularly sensitive to the choice of initial parameters so it seems like a big success.