transitioning away from lbfgsb
fit_templates and fit_templates_mdf are both using log transforms now, first with a prior for regularization and convergence, then they are re-optimized without the prior to find the maximum likelihood. By using bfgs with also get the inverse hessian estimate, which turns out to be fully sufficient to replicate the results of hmc_sample and hmc_sample_mdf, at least when they are using the prior as well. Initial tests indicate it is possible to run the HMC without the prior, but the trajectories tend to be divergent and the HMC step size is required to be very low, which makes it inefficient. The new implementations for fit_templates and fit_templates_mdf now contain information necessary to derive the same information as can be determined from the HMC, via the inverse Hessian approximation.