continuing example -- found bug??
Think there is an error in how I'm treating the logAge vector when computing SFRs. Im assuming the most recent template (e.g., logAge=6.6) should stretch all the way to present day when in reality I should just require a `length(templates)+1` vector for logAge that fully defines the bin edges for the templates. If uniformly spaced in logAge, we should respect that