add max_logAge to set final bin width
Altered `construct_x0`, `calculate_cum_sfr`, and `construct_x0_mdf` to have a `max_logAge` argument that sets the final bin width. I.e., if the unique entries in logAge are [6.6, 6.7, 6.8], we treat those as left-bins for defining star formation rates, and you would probably want to set `max_logAge=6.9` so that the bins have equal spacing. In the case that your isochrone grid is dense in logAge and goes to very old ages (e.g., logAge of 10.10) you could set an upper bound of the Hubble time `log10(13.7e9)`. This unfortunately DOES NOT fix our bias towards greater SFRs at more recent times evident in fitting1.ipynb. Going to have to go back to the prior derivation and see if there's a problem there -- I think it is very likely related to the priors and the logarithmic transform I'm doing on the fitting coefficients to guarantee they are positive.