initial commit with new binary model
Implements a binary model `BinaryMassRatio` that takes in a system mass and portions it out to a primary and secondary according to a provided mass ratio distribution. Also renamed `Binaries` to `RandomBinaryPairs`. The current `sample_binary` API is suboptimal for this application; think I will attempt to revise it tomorrow.