tried to derive more accurate \sigma, failed
Trying to return a more useful estimate of the 1-sigma uncertainties from the sqrt transform MLE but my new derivation returns uncertainties that are much too high. If off-diagonal elements are important in the covariance matrix than a standard deviation formulation will never be very accurate. Leaving for now, but in the future we may simply want to get rid of \sigma and force people to calculate it themselves from samples constructed via `rand`.