New tests for `fixed_lamr`, begin`fixed_amr`
Tests are set up so that when we switch the `fixed_lamr` implementation we should be able to verify that the new implementation works. Going to create a new `fixed_amr` method that accepts a relative weight vector which contains the weights due to the assumed metallicity-age relation. This is a generalization of the current `fixed_lamr` which assumes one particular form for this AMR.