Remove `test/Project.toml` and add compat to main `Project.toml`
While the `test/Project.toml` environment is a valid way to organize a test environment, because this project has more dependencies than my previous projects I don't want to have to manage the compatibility entries in the main project and the test project separately. By moving to a unified main `Project.toml`, we can control all the versioning compatibility in one file and include extra packages for testing in the [extras] section.