Start working on logarithmic AMR
`fixed_log_amr` is implemented and seems to work, example of the AMR is provided in `examples/log_amr/log_amr_example.jl`. Additionally `calculate_coeffs_logamr` in `log_amr.jl` has a docstring but it contains references to methods that are not programmed yet, namely methods that will fit the AMR coefficients. These would be nice to have but will take a longer time to develop. Tomorrow want to try using `fixed_log_amr` in production. Should provide an improvement over the current hack of using `fixed_linear_amr` with input Z, because that method requires a very low `\sigma` value in Z rather than a spread in [M/H] dex which is more physical.