Continue optimizations in `fitting_base.jl`
Add call signatures supporting the flattened model format we added to `composite!`. Of particular importance is `∇loglikelihood!` which has been updated in a similar way to `composite!` to use a matrix-vector product computed with `LinearAlgebra.mul!`. For small inputs this is not a huge performance improvement, but seeing up to 40% performance improvement with single-threaded OpenBLAS for large input model lists. Presumably this improves even more with vendor BLAS (e.g., MKL or Apple Accelerate). Tests covering these new signatures have been added. Should be time to start introducing the new signatures into the fitting functions and ensuring that their results are robust.