Major: Begin implementing `fit_templates_log_amr`
This is a major commit beginning the work on the methods to fit logarithmic age-metallicity relations. Prior to this I only had a fixed fitting method `fixed_log_amr`. This seems to be a better model than the linear [M/H] model, so I'm writing full fitting functions for the logarithmic [M/H] model. Gradient calculation in `fitting/hierarchical/log_amr/fg_log_amr!` is consistent with ForwardDiff result; derivation and implementation both seem correct. Next step is write the optimization function that will use the gradient to perform the optimization.

Implementing this required revising the metallicity utilities, `MH_from_Z`, `Z_from_MH`, `X_from_Z`, `Y_from_Z`, and includes the addition of `dMH_dZ`. In particular, the slope used in `Y_from_Z` is now an additional argument rather than hard-coded to a single value. This change has been passed through the other functions that call `Y_from_Z` to support this new argument.

This commit also reorganizes the docs for the "Deriving Star Formation Histories..." section. We are using the `pages` keyword argument to `Documenter.makedocs` to nest more headings under this section and provide better separation of the different fitting approaches. New documentation has been begun for the logarithmic AMR implementation. We are making a figure to display in the documentation on the fly using Plots.jl, which is now a dependency in the `docs/Project.toml`. I additionally added tests `test/fitting/linear_amr_test.jl` and `test/fitting/log_amr_test.jl` which are structured as independently runable files which are included in the main `runtests.jl` using `SafeTestsets.jl`, based on the implementation in `OrdinaryDiffEq.jl`. This should make these tests more easily separable, alterable, and maintainable. Would like to migrate the majority of the tests to this format in the future.