Prep for change to log AMR model
Initial tests with fitting the logarithmic AMR model indicate that the issue I predicted with the definition of alpha and beta are indeed problems. Because beta is defined at present-day, it is very easy for the slope alpha to become too steep so that Z becomes negative at earlier times, which is unallowed by the model. We can prevent this by redefining beta to be Z at the earliest possible time (`max_logAge` used in some other functions), then alpha is a positive slope such that`Z_t = (exp10(max_logAge) - t) / 1e9 * alpha + beta` where `t` is the lookback time in Gyr. Under this reparameterization, alpha beta and sigma must all be positive such that ALL the fitting variables must be positive, simplifying our transformations. This is, however, annoying as it is the opposite parameterization from the existing `linear_amr` work which defines beta at present-day. If we reparameterize here, we will need to alter the other functions (e.g., `fixed_log_amr`) to use this new parameterization as well.