Start work on adding back covariant kernels
For the separable kernel, we can easily support different bin widths in x and y when computing the integral of the kernel across each bin. In the covariant case, the integral across the pixel has different widths in magnitude space which causes trouble with the covariance matrix. Might need a more general form of the `...halfpix` integrals in order to support this use case.