first covariant kernel looks good
main difficulty appears to be the conversion from magnitude space (which is more intuitive and probably simpler) to Hess diagram pixel space (which is how I'm currently defining the kernels). The pixel space implementation is more efficient because you don't have to check bin sizes and such, but requires special care when setting up the kernel and afterwards as well. For example, you need to renormalize the output matrix according to the `ybin_factor` to restore the sum-to-1 normalization.