break code out into generic `gaussian_int_general`
For `Float32/64` we can use `VectorizationBase.verf` to get a ~2x speedup in `addstar!` which should be worth it as `VectorizationBase` is a dependency of `LoopVectorization` already. To support this specialization, I was able to break the main `erf` calls out into `gaussian_int_general` so that the kernel `evaluate` methods for both `GaussianPSFAsymmetric` and `GaussianPSFCovariant` now call into this single generic function. We are still keeping a fallback for general `Real` inputs to `SpecialFunctions.erf` which supports more datatypes. Also added some docstrings and cleaned up some code syntax throughout.