Add `@turbo` to `addstar!` for `PixelSpaceKernel`
Additional 2x speed improvement due to better simd. Requires ensuring that the loop indices are all inbounds *prior* to starting the loop, where the old implementation was lazy and looped over all defined indices with explicit `checkbounds()` inside the loop. `LoopVectorization.jl` and `VectorizationBase.jl` are at risk of being deprecated in Julia 1.11, but using them together in this application gives 4x speedup for the hottest loop in the template creation process, so I think this is worth doing for now.