Increase resolution in `smooth_template.jl`
At lower Hess diagram resolutions like `size(edges) = (75, 75)` there were significant artifacts in the template model with the new covariant kernel. Doubling the y-resolution to `length=150` gives a much better model. Not obvious why this is the case, as the integration over the pixels should make the solution resolution agnostic. Might have something to do with not resolving the skewed kernel shape if the grid is too coarse? Worth more investigation.