`calculate_edges` checks ranges for input `edges`
Previously if you passed a direct `edges` argument to `calculate_edges` it would return it with no checks. Downstream functions that call this method now assume the output will be a `Tuple{<:AbstractRange, <:AbstractRange}`, so we have to check the input type and error if it is incorrect. This commit also cleans the docs for `bin_cmd`.