increase `size` for `GaussianPSFCovariant`
old size was mostly adequate, but x-axis width was a bit low such that I could see the truncation of the kernel in the plots I'm making for the paper. Since the methods are well optimized, we can afford to increase the size of the kernel; making a single template still takes ~3 ms so still quite good.