correct initial mass pdf norm in `binary_hess`
In the binary case, we need to add an additional factor of `integral(pdf(imf, m), extrema(m_ini)...)` in the weights to get the scaling correct. We could use `quadgk` for this but it seems like overkill; importing `Trapz.trapz` instead. We also need to know the fraction of total stellar mass in binary systems; new method `binary_mass_fraction(::AbstractBinaryModel, imf)` for this purpose. Implemented only for `RandomBinaryPairs` so far.