fix `binary_hess` weighting for `RandomBinaryPairs`
I kept trying to fix the problem by altering `binary_hess` but that actually wasn't where the problem was. Consider that there will be an excess of single stars at present-day due to stars born in binary pairs that have died and do not appear in the isochrone. We must account for this additional factor, which is the integral of the imf over the range of m_ini. This should be done in the co-addition stage in `partial_cmd_smooth`.