Add support for `RandomBinaryPairs` binary model to `partial_cmd_smooth` (#42)
Major update that implements the RandomBinaryPairs binary model in the smooth template creation process of partial_cmd_smooth. The binaries branch has some additional code where I tried to add BinaryMassRatio but that turned out to be tricky so giving up on trying to get it into the initial release.