Only fit free variables -- fix sampling methods
To address #53 before generalizing the new framework to AMRs as well, the `logdensity_and_gradient` method nowexpects input vectors to have length equal to the number of free parameters (fixed parameters are no longer included). We support this by using a temporary array `G2`, which has a length equal to the number of all parameters (fixed and free) which goes into `fg_mzr!`; then the gradient values for the *free* parameters are copied from `G2` and written into `G`, the gradient array that Optim.jl and DynamicHMC.jl see. This means that the minimizers coming out of the BFGS optimization do not include fixed variables, nor do the inverse Hessians. This is good for DynamicHMC -- the efficiency problem noted in #53 is fixed by this change. It does mean we need to augment all the random sampling methods with the fixed parameters after the fact, but this has been taken care of. The code is now more complicated (it can be tricky to track all the indexing operations), but the core functionality we wanted is there and performance is good. I have added tests to check this functionality, and all previous tests pass as well. Tomorrow I will test that things still work on real data before PRing onto MZR branch.