Stabilizing `LogarithmicAMR` sfh solves
During linesearch for BFGS it is possible for the algorithm to try a combination of slope and intercept that results in negative mass fractions, which causes `log10` to throw errors and interrupt the algorithm. We can avoid this scenario by returning `NaN` out of `MH_from_Z` instead of allowing `log10` to error.