Unify `process_ASTs` implementation, remove exts
The main problem that kept us from using the same implementation for both DataFrames and TypedTables was that they do not return the same types when iterated with Base.`eachrow`. The correct way to iterate rows for tables is to use `Tables.rows`. Making this change allows us to remove the package extensions for TypedTables and DataFrames, though we continue to test against both.